src/AnalysisLayer/Vision/intelligencelayer/requirements.txt
#! Python3.7

## CodeProject.AI - make this a "-r CodeProject.AI.requirements.txt"
requests         # Installing Requests, the HTTP library
Pillow           # Installing Pillow, a Python Image Library

## Needed, but not used everywhere.
THOP             # Installing THOP (PyTorch-OpCounter) for measuring performance
ONNX             # Installing ONNX, the Open Neural Network Exchange library
Pandas           # Installing Pandas, a data analysis / data manipulation tool
CoreMLTools      # Installing CoreMLTools, for working with .mlmodel format models

## General libraries
ONNXRuntime      # Installing ONNX runtime, the scoring engine for ONNX models
SciPy            # Installing SciPy, a library for mathematics, science, and engineering
PyYAML           # Installing PyYAML, a library for reading configuration files

## Specific versions that match the models we're using. 
--find-links https://download.pytorch.org/whl/torch_stable.html
Torch==1.6.0+cpu         # Installing PyTorch, for Tensor computation and Deep neural networks
--find-links https://download.pytorch.org/whl/torch_stable.html
TorchVision==0.7.0+cpu  # Installing TorchVision, for Computer Vision based AI

## No longer needed
## Seaborn       # library for making statistical graphics
## wMatplotlib
## OpenCV-Python
## Cython
## TDQM 
## Tensorboard